Completed User groups
Added user groups to the forms and collection. I made the columns color
coded according to the group by calling a function in the html homepage
that users a helper function in the js page that returns the class name
of the group. The css file then colors the rows according to the group
in the class name.